refactor: deprecate StoryValidator
