+++ botfront/cypress/integration/stories/extract_domain.spec.js
@@ -2,7 +2,15 @@
 import { safeLoad } from 'js-yaml';
 import { isEqual } from 'lodash';
 import gold2 from '../../fixtures/story_exceptions.json';
-import { extractDomain, StoryValidator } from '../../../imports/lib/story_validation';
+import { StoryController, extractDomain } from '../../../imports/lib/story_controller';
+
+const slots = [
+    {
+        name: 'feedback_value',
+        type: 'categorical',
+        categories: ['positive', 'negative'],
+    },
+];
 
 describe('extract domain from storyfile fixtures', function() {
     before(function() {
@@ -14,13 +22,13 @@ describe('extract domain from storyfile fixtures', function() {
 
     it('should output yaml matching the gold', function() {
         const gold = safeLoad(this.gold);
-        const domain = safeLoad(extractDomain(this.stories.split('\n\n')));
+        const domain = safeLoad(extractDomain(this.stories.split('\n\n'), slots));
         expect(isEqual(domain, gold)).to.be.equal(true);
     });
 
     it('should output exceptions matching the gold', function() {
-        const val = new StoryValidator(this.stories2);
-        val.validateStories();
+        const val = new StoryController(this.stories2, slots);
+        val.validateStory();
         const exceptions = val.exceptions.map(exception => ({
             line: exception.line,
             code: exception.code,
