+++ botfront/cypress/integration/stories/stories.spec.js
@@ -2,7 +2,6 @@
 
 const storyGroupOne = 'storyGroupOne';
 const storyGroupTwo = 'storyGroupTwo';
-const initialText = '';
 const testText = '* my_intent';
 
 describe('stories', function() {
@@ -32,7 +31,6 @@ describe('stories', function() {
         clickStoryGroup(storyGroupOne);
         clickStoryGroup(storyGroupOne);
         cy.dataCy('story-title').should('have.value', storyGroupOne);
-        cy.dataCy('story-editor').contains(initialText);
         cy.dataCy('move-story').click({ force: true });
         cy.dataCy('move-story-dropdown').click({ force: true });
         cy.dataCy('move-story-dropdown')
@@ -88,8 +86,6 @@ describe('stories', function() {
             .type('Group 2 {enter}');
         // Add a story to group 1
         clickStoryGroup('Group 1');
-        // Verify that default story is present
-        cy.dataCy('story-editor').contains(initialText);
         // Edit story content
         cy.dataCy('story-editor')
             .get('textarea')
@@ -100,7 +96,7 @@ describe('stories', function() {
         // Go back to group 1
         clickStoryGroup('Group 1');
         // We should find the story content we saved...
-        cy.contains(`${initialText}xxx`).should('exist');
+        cy.contains('xxx').should('exist');
     });
 
     it('should be able to duplicate a story', function() {
@@ -110,7 +106,6 @@ describe('stories', function() {
             .find('input')
             .type(`${storyGroupOne}{enter}`);
         cy.wait(200);
-        cy.dataCy('story-editor').contains(initialText);
         cy.dataCy('duplicate-story').click({ force: true });
         cy.dataCy('story-editor').should('have.lengthOf', 2);
         cy.dataCy('delete-story')
@@ -155,7 +150,6 @@ describe('stories', function() {
             .find('input')
             .type(`${storyGroupOne}{enter}`);
         clickStoryGroup(storyGroupOne);
-        cy.dataCy('story-editor').contains(initialText);
         cy.dataCy('story-editor')
             .get('textarea')
             .type(`{selectall}{backspace}${testText}`, { force: true });
